cfo-cli-pkg/pyproject.toml
[project]
name = "cfo-cli"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "click>=8.1,<9",
  "httpx>=0.28,<1",
  "pydantic>=2.0,<3",
  "pyyaml>=6.0,<7",
]

[project.scripts]
cfo-cli = "cfo_cli.main:cli"
